requirements.txt
accelerate==0.22.0
langchain==0.0.283
sentence-transformers==2.2.2
transformers==4.33.1
torch
bitsandbytes==0.41.1
chromadb==0.4.9
html2text==2020.1.16
streamlit==1.26.0
python-dotenv==1.0.0

